iot_blockchain_project/requirements.txt
paho-mqtt==1.6.1
pycryptodome==3.20.0
matplotlib
networkx==3.2.1
psutil==5.9.7